+++ crates/wdl-analysis/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Fixed
 
 * Placeholder options are now statically type checked ([#345](https://github.com/stjude-rust-labs/wdl/pull/345)).
+* Prevent lsp crash due to panic in single file analysis ([#431](https://github.com/stjude-rust-labs/wdl/pull/431)).
 
 ## 0.7.0 - 04-01-2025